feat: eviction + geo-nearest n fix
1) Fixes a bug where we could end up with extra results when using
`geo-nearest` across attributes... Namely, we couldn't rely on the
`not=` on the when to filter down to the appropriate number of elements
when joining across attributes (since it wouldn't be equal, and so the
`inc` would give us an extra value). We now bound it with a `take`.

2) Adds support for evicting values. The only thing interesting here is
that we can't rely on passing in the geometry directly, instead we need
to query to get the actual instance that's in the tree.